Fixed animethemes data caching.
It now checks every anime respectively, so it doesn't need to fetch unneeded data.

Next step is to cache MAL and Anilist results, that would mean normalizing animelist getting, which I wanted to do anyways.
I will also be creating a better way to deal with honorifics in anime titles, so it's possible to find the missing anime.